refactor!: only support winston3
BREAKING CHANGE: Drop support for winston2 as a semver major change.

Ref: https://github.com/googleapis/nodejs-logging-winston/pull/278
