+++ src/types/core.d.ts
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-import * as TransportStream from "winston-transport";
+import * as TransportStream from 'winston-transport';
+import {GoogleAuth} from 'google-auth-library';
 
 
 export interface Options {
@@ -168,6 +169,7 @@ interface StackdriverOtherFunctions {
        callback?:
            (err: Error, apiResponse: {}) => void) => Promise<LogWriteResponse>;
   entry: (metadata: {}, data: {}|string) => StackdriverEntry;
+  logging: { auth: GoogleAuth };
 }
 
 export type StackdriverLog = StackdriverLogFunctions&StackdriverOtherFunctions;
