--- test/winston-2/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "winston-holder",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "winston": "^2.2.0"
-  }
-}
